moved otel.py back into zen_server module
