+++ src/zenml/zen_server/zen_server_api.py
@@ -45,13 +45,13 @@ from zenml.constants import (
     READY,
 )
 from zenml.enums import AuthScheme
-from zenml.otel import configure_otel, otel_span, shutdown_otel
 from zenml.service_connectors.service_connector_registry import (
     service_connector_registry,
 )
 from zenml.zen_server.cloud_utils import send_pro_workspace_status_update
 from zenml.zen_server.exceptions import error_detail
 from zenml.zen_server.middleware import add_middlewares
+from zenml.zen_server.otel import configure_otel, otel_span, shutdown_otel
 from zenml.zen_server.routers import (
     artifact_endpoint,
     artifact_version_endpoints,
